+++ CHANGELOG.rst
@@ -3,6 +3,7 @@ Release 0.X.0 (unreleased)
 
 * Add diff command for svn projects (#24)
 * Also add binary files as part of generated patch (#251)
+* Create diff on working copy instead of current revision (#254)
 
 Release 0.4.0 (released 2021-11-26)
 ===================================